addressing KS's comments round 4
