+++ tests/python/operators/fermion_operator_test.py
@@ -643,3 +643,12 @@ def test_mapping_methods():
         ((ffsim.cre_b(1), ffsim.des_b(2)), -0.5j),
         ((ffsim.cre_b(2), ffsim.des_b(1)), 1 - 0.5j),
     }
+
+def test_trace():
+    norb = 50
+    nelec = (3, 3)
+    rng = np.random.default_rng(12345)
+    ham = ffsim.random.random_diagonal_coulomb_hamiltonian(norb, real=True, seed=rng)
+    t1 = ffsim.trace(ffsim.fermion_operator(ham), norb=norb, nelec=nelec)
+    t2 = ffsim.trace(ham, norb=norb, nelec=nelec)
+    np.testing.assert_allclose(t1, t2)
